Heightmap: Properly implement reference tracking unload approach
SurfaceTrackerSection was split into a base class SurfaceTrackerNode, and two
derived classes: SurfaceTrackerBranch, SurfaceTrackerLeaf, which splits their
implementation making it far simpler.

SurfaceTrackerBranches have a requiredChildren value, which for scale 1 branches
is the number of their children (leaves) with a cube loaded. Any node with
requiredChildren will be required by its parent.
If the requiredChildren is 0, all children can be unloaded.

Co-authored-by: Christopher Cyclonit Klinge <christ.klinge@web.de>
Co-authored-by: Bartosz Skrzypczak <barteks2x@gmail.com>